Embed HoverLayer in ListMode instead of it always being active
