+++ src/lib/draw/HoverLayer.svelte
@@ -8,20 +8,14 @@
     isPolygon,
     layerId,
   } from "lib/maplibre";
-  import { gjScheme, mode, sidebarHover } from "stores";
+  import { gjScheme, sidebarHover } from "stores";
   import { CircleLayer, GeoJSON, LineLayer } from "svelte-maplibre";
 
   // Use a layer that only ever has zero or one features for hovering.
-  let gj: GeoJSONType = emptyGeojson();
-
-  $: {
-    let id = $mode.mode == "list" ? $sidebarHover : null;
-    if (id == null) {
-      gj = emptyGeojson();
-    } else {
-      gj = $gjScheme.features.find((f) => f.id == id)!;
-    }
-  }
+  $: gj =
+    $sidebarHover == null
+      ? emptyGeojson()
+      : $gjScheme.features.find((f) => f.id == $sidebarHover)!;
 </script>
 
 <GeoJSON data={gj}>
